Split DataAdapter in multiple smaller classes
